// app/src/api/models/TokenDto_Output.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TokenDto_Output = {
  symbol: string
  market_id: number
  price: string
  max_price?: string | null
  min_price?: string | null
  status?: string | null
  min_base_amount?: string | null
  min_quote_amount?: string | null
  size_decimals: number
  price_decimals: number
}